ci: create a PR for bumping version
pushing directly to main strategy won't work thanks to branch protection
rules.

Instead, use a separate action that opens a PR bumping version and
updating changelog.
After merge, a release will be created.

The distinction on what to do is
solely based on the latest commit message: if it starts with bump,
create a release; otherwise, open a PR bumping. This adds some
flexibility for developers manually forcing a version to a certain
number when desired.

Other changes:
- replace GA version of commitizen with a custom action to avoid
creating tags (which should be done when creating a new release);
- replace commitsar with commitizen for checking commits.

fix: DISCOVERY-291